+++ packages/pangea-sdk/pangea/asyncio/services/authn.py
@@ -25,14 +25,14 @@ class AuthNAsync(ServiceBaseAsync):
         import os
 
         # Pangea SDK
+        from pangea.asyncio.services import AuthNAsync
         from pangea.config import PangeaConfig
-        from pangea.services import AuthN
 
         PANGEA_TOKEN = os.getenv("PANGEA_AUTHN_TOKEN")
         authn_config = PangeaConfig(domain="pangea.cloud")
 
         # Setup Pangea AuthN service
-        authn = AuthN(token=PANGEA_TOKEN, config=authn_config)
+        authn = AuthNAsync(token=PANGEA_TOKEN, config=authn_config)
     """
 
     service_name = SERVICE_NAME
@@ -450,6 +450,8 @@ class AuthNAsync(ServiceBaseAsync):
             self,
             email: str,
             profile: m.Profile,
+            *,
+            username: str | None = None,
         ) -> PangeaResponse[m.UserCreateResult]:
             """
             Create User
@@ -459,8 +461,9 @@ class AuthNAsync(ServiceBaseAsync):
             OperationId: authn_post_v2_user_create
 
             Args:
-                email (str): An email address
-                profile (m.Profile): A user profile as a collection of string properties
+                email: An email address.
+                profile: A user profile as a collection of string properties.
+                username: A username.
 
             Returns:
                 A PangeaResponse with a user and its information in the response.result field.
@@ -479,11 +482,12 @@ class AuthNAsync(ServiceBaseAsync):
             input = m.UserCreateRequest(
                 email=email,
                 profile=profile,
+                username=username,
             )
             return await self.request.post("v2/user/create", m.UserCreateResult, data=input.dict(exclude_none=True))
 
         async def delete(
-            self, email: Optional[str] = None, id: Optional[str] = None
+            self, email: str | None = None, id: str | None = None, *, username: str | None = None
         ) -> PangeaResponse[m.UserDeleteResult]:
             """
             Delete User
@@ -493,8 +497,9 @@ class AuthNAsync(ServiceBaseAsync):
             OperationId: authn_post_v2_user_delete
 
             Args:
-                email (str, optional): An email address
-                id (str, optional): The id of a user or a service
+                email: An email address.
+                id: The id of a user or a service.
+                username: A username.
 
             Returns:
                 A PangeaResponse with an empty object in the response.result field.
@@ -502,7 +507,7 @@ class AuthNAsync(ServiceBaseAsync):
             Examples:
                 authn.user.delete(email="example@example.com")
             """
-            input = m.UserDeleteRequest(email=email, id=id)
+            input = m.UserDeleteRequest(email=email, id=id, username=username)
             return await self.request.post("v2/user/delete", m.UserDeleteResult, data=input.dict(exclude_none=True))
 
         async def invite(
@@ -549,8 +554,10 @@ class AuthNAsync(ServiceBaseAsync):
         async def update(
             self,
             disabled: bool,
-            id: Optional[str] = None,
-            email: Optional[str] = None,
+            id: str | None = None,
+            email: str | None = None,
+            *,
+            username: str | None = None,
         ) -> PangeaResponse[m.UserUpdateResult]:
             """
             Update user's settings
@@ -560,10 +567,11 @@ class AuthNAsync(ServiceBaseAsync):
             OperationId: authn_post_v2_user_update
 
             Args:
-                disabled (bool): New disabled value.
+                disabled: New disabled value.
                     Disabling a user account will prevent them from logging in.
-                id (str, optional): The identity of a user or a service
-                email (str, optional): An email address
+                id: The identity of a user or a service.
+                email: An email address.
+                username: A username.
 
             Returns:
                 A PangeaResponse with a user and its information in the response.result field.
@@ -580,6 +588,7 @@ class AuthNAsync(ServiceBaseAsync):
                 id=id,
                 email=email,
                 disabled=disabled,
+                username=username,
             )
 
             return await self.request.post("v2/user/update", m.UserUpdateResult, data=input.dict(exclude_none=True))
@@ -710,7 +719,12 @@ class AuthNAsync(ServiceBaseAsync):
                 super().__init__(token, config, logger_name=logger_name)
 
             async def delete(
-                self, authenticator_id: str, id: Optional[str] = None, email: Optional[str] = None
+                self,
+                authenticator_id: str,
+                id: str | None = None,
+                email: str | None = None,
+                *,
+                username: str | None = None,
             ) -> PangeaResponse[m.UserAuthenticatorsDeleteResult]:
                 """
                 Delete user authenticator
@@ -720,9 +734,10 @@ class AuthNAsync(ServiceBaseAsync):
                 OperationId: authn_post_v2_user_authenticators_delete
 
                 Args:
-                    authenticator_id (str): An ID for an authenticator
-                    id (str, optional): The identity of a user or a service
-                    email (str, optional): An email address
+                    authenticator_id: An ID for an authenticator.
+                    id: The identity of a user or a service.
+                    email: An email address.
+                    username: A username.
 
                 Returns:
                     A PangeaResponse with an empty object in the response.result field.
@@ -733,7 +748,9 @@ class AuthNAsync(ServiceBaseAsync):
                         id="pui_xpkhwpnz2cmegsws737xbsqnmnuwtbm5",
                     )
                 """
-                input = m.UserAuthenticatorsDeleteRequest(authenticator_id=authenticator_id, email=email, id=id)
+                input = m.UserAuthenticatorsDeleteRequest(
+                    authenticator_id=authenticator_id, email=email, id=id, username=username
+                )
                 return await self.request.post(
                     "v2/user/authenticators/delete",
                     m.UserAuthenticatorsDeleteResult,
@@ -741,7 +758,7 @@ class AuthNAsync(ServiceBaseAsync):
                 )
 
             async def list(
-                self, email: Optional[str] = None, id: Optional[str] = None
+                self, email: str | None = None, id: str | None = None, *, username: str | None = None
             ) -> PangeaResponse[m.UserAuthenticatorsListResult]:
                 """
                 Get user authenticators
@@ -751,8 +768,9 @@ class AuthNAsync(ServiceBaseAsync):
                 OperationId: authn_post_v2_user_authenticators_list
 
                 Args:
-                    email (str, optional): An email address
-                    id (str, optional): The identity of a user or a service
+                    email: An email address.
+                    id: The identity of a user or a service.
+                    username: A username.
 
                 Returns:
                     A PangeaResponse with a list of authenticators in the response.result field.
@@ -764,7 +782,7 @@ class AuthNAsync(ServiceBaseAsync):
                         id="pui_xpkhwpnz2cmegsws737xbsqnmnuwtbm5",
                     )
                 """
-                input = m.UserAuthenticatorsListRequest(email=email, id=id)
+                input = m.UserAuthenticatorsListRequest(email=email, id=id, username=username)
                 return await self.request.post(
                     "v2/user/authenticators/list", m.UserAuthenticatorsListResult, data=input.dict(exclude_none=True)
                 )
@@ -781,7 +799,7 @@ class AuthNAsync(ServiceBaseAsync):
                 super().__init__(token, config, logger_name=logger_name)
 
             async def get(
-                self, id: Optional[str] = None, email: Optional[str] = None
+                self, id: str | None = None, email: str | None = None, *, username: str | None = None
             ) -> PangeaResponse[m.UserProfileGetResult]:
                 """
                 Get user
@@ -791,8 +809,9 @@ class AuthNAsync(ServiceBaseAsync):
                 OperationId: authn_post_v2_user_profile_get
 
                 Args:
-                    id (str, optional): The identity of a user or a service
-                    email (str, optional): An email address
+                    id: The identity of a user or a service.
+                    email: An email address.
+                    username: A username.
 
                 Returns:
                     A PangeaResponse with a user and its information in the response.result field.
@@ -804,7 +823,7 @@ class AuthNAsync(ServiceBaseAsync):
                         email="joe.user@email.com",
                     )
                 """
-                input = m.UserProfileGetRequest(id=id, email=email)
+                input = m.UserProfileGetRequest(id=id, email=email, username=username)
                 return await self.request.post(
                     "v2/user/profile/get", m.UserProfileGetResult, data=input.dict(exclude_none=True)
                 )
@@ -812,8 +831,10 @@ class AuthNAsync(ServiceBaseAsync):
             async def update(
                 self,
                 profile: m.Profile,
-                id: Optional[str] = None,
-                email: Optional[str] = None,
+                id: str | None = None,
+                email: str | None = None,
+                *,
+                username: str | None = None,
             ) -> PangeaResponse[m.UserProfileUpdateResult]:
                 """
                 Update user
@@ -823,9 +844,10 @@ class AuthNAsync(ServiceBaseAsync):
                 OperationId: authn_post_v2_user_profile_update
 
                 Args:
-                    profile (m.Profile): Updates to a user profile
-                    id (str, optional): The identity of a user or a service
-                    email (str, optional): An email address
+                    profile: Updates to a user profile.
+                    id: The identity of a user or a service.
+                    email: An email address.
+                    username: A username.
 
                 Returns:
                     A PangeaResponse with a user and its information in the response.result field.
@@ -844,6 +866,7 @@ class AuthNAsync(ServiceBaseAsync):
                     id=id,
                     email=email,
                     profile=profile,
+                    username=username,
                 )
                 return await self.request.post(
                     "v2/user/profile/update", m.UserProfileUpdateResult, data=input.dict(exclude_none=True)